# public/doc_swagger/openapi.yaml
openapi: 3.0.0
info:
  title: 'Api oAuth 2.0'
  description: 'Esta é a documentação base da sua API. Ela contém todas as informações necessárias para que você, sua equipe ou um desenvolvedor possam integrar o seu sistema com outros sistemas, sites, portais ou aplicativos móveis (iOS/Android)'
  version: 2.0.1
paths:
...
  '/api/permissions/user/{id}':
    get:
      tags:
        - 'Permissão de Acesso'
      summary: 'Listar permissões do usuário'
      operationId: a7b4ac3ed70b30eac404e92e4ebd0de6
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: int
        -
          name: Authorization
          in: header
          required: true
          schema:
            description: 'Bearer token para autorização'
            type: string
            default: 'Bearer <token>'
      responses:
        '200':
          description: ''
    put:
      tags:
        - 'Permissão de Acesso'
      summary: 'Alterar permissões do usuário'
      operationId: 76492792cd15ef30d726f656dc0321d5
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: string
        -
          name: Authorization
          in: header
          required: true
          schema:
            description: 'Bearer token para autorização'
            type: string
            default: 'Bearer <token>'
      requestBody:
        description: 'Input data format'
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                properties:
                  role: { description: 'Função/Role do Controller', type: string }
                  authorized: { description: 'Indica se está autorizado ou não', type: boolean }
                type: object
      responses:
        '200':
          description: ''
  '/api/permissions/profile/{profile_id}':
    get:
      tags:
        - 'Permissão de Acesso'
      summary: 'Listar permissões do perfil de acesso'
      operationId: bd37e166639f813824a5d73ac2001b10
      parameters:
        -
          name: profile_id
          in: path
          required: true
          schema:
            type: int
        -
          name: Authorization
          in: header
          required: true
          schema:
            description: 'Bearer token para autorização'
            type: string
            default: 'Bearer <token>'
      responses:
        '200':
          description: ''
    put:
      tags:
        - 'Permissão de Acesso'
      summary: 'Alterar permissões do perfil de acesso'
      operationId: bf7b1b37fb76ef1a2df4a6ccb854e493
      parameters:
        -
          name: profile_id
          in: path
          required: true
          schema:
            type: int
        -
          name: Authorization
          in: header
          required: true
          schema:
            description: 'Bearer token para autorização'
            type: string
            default: 'Bearer <token>'
      requestBody:
        description: 'Input data format'
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                properties:
                  role: { description: 'Função/Role do Controller', type: string }
                  authorized: { description: 'Indica se está autorizado ou não', type: boolean }
                type: object
      responses:
        '200':
          description: ''
...